backend/requirements.txt
fastapi==0.112.1
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.32
langchain_openai==0.1.21
langchain_text_splitters==0.2.2
matplotlib==3.6.2
pandas==2.2.2
python-dotenv==1.0.1
requests==2.28.1
SQLAlchemy==2.0.32
tiktoken==0.7.0
uvicorn==0.30.5
faiss-cpu==1.8.0